Add files via upload
In this project, I will be performing an unsupervised clustering of data on the customer's records from a groceries firm's database.

Customer segmentation is the practice of separating customers into groups that reflect similarities among customers in each cluster. 

I will divide customers into segments to optimize the significance of each customer to the business. Modify products according to the distinct needs and behaviors of the customers. It also helps the business to cater to the concerns of different types of customers.

In this problem, there are many factors on the basis of which the final classification will be done. 

These factors are basically attributes or features. The higher the number of features, the harder it is to work with them. Many of these features are correlated, and hence redundant.

This is why I will be performing dimensionality reduction on the selected features before putting them through a classifier.